+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 
 - **feat**: Add support for saving unsaved changes in the editor before
   compilation.
+- **fix**: The "file not found" warning no longer appears when the current
+  source file does not have a problem.
+- **feat**: Add support for SPJ using testlib.
 
 ## 0.0.2